feat: clean up adapters
