+++ lib/lutaml/model.rb
@@ -1,11 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "model/version"
-
-# lib/lutaml/model.rb
-require "nokogiri"
-require "json"
-require "yaml"
 require_relative "model/type"
 require_relative "model/serializable"
 require_relative "model/json_adapter"
