+++ lib/lutaml/model/json_adapter/json_document.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+require_relative "json_object"
+
+module Lutaml
+  module Model
+    module JsonAdapter
+      # Base class for JSON documents
+      class JsonDocument < JsonObject
+        def self.parse(json)
+          raise NotImplementedError, "Subclasses must implement `parse`."
+        end
+
+        def to_json(*args)
+          raise NotImplementedError, "Subclasses must implement `to_json`."
+        end
+      end
+    end
+  end
+end
