+++ lib/lutaml/model/toml_adapter/toml_document.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require_relative "toml_object"
+
+module Lutaml
+  module Model
+    module TomlAdapter
+
+      # Base class for TOML documents
+      class TomlDocument < TomlObject
+        def self.parse(toml)
+          raise NotImplementedError, "Subclasses must implement `parse`."
+        end
+
+        def to_toml(*args)
+          raise NotImplementedError, "Subclasses must implement `to_toml`."
+        end
+      end
+    end
+  end
+end
