+++ lib/lutaml/model/yaml_adapter/yaml_document.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+require_relative "yaml_object"
+
+module Lutaml
+  module Model
+    module YamlAdapter
+      # Base class for YAML documents
+      class YamlDocument < YamlObject
+        def self.parse(yaml)
+          raise NotImplementedError, "Subclasses must implement `parse`."
+        end
+
+        def to_yaml(*args)
+          raise NotImplementedError, "Subclasses must implement `to_yaml`."
+        end
+      end
+    end
+  end
+end
