+++ spec/lutaml/model/toml_adapter_spec.rb
@@ -1,4 +1,3 @@
-# spec/lutaml/model/toml_adapter_spec.rb
 require "spec_helper"
 require "lutaml/model/toml_adapter/toml_rb_adapter"
 require "lutaml/model/toml_adapter/tomlib_adapter"
@@ -9,9 +8,9 @@ RSpec.shared_examples "a TOML adapter" do |adapter_class|
   let(:model) { SampleModel.new(attributes) }
 
   let(:expected_toml) do
-    if adapter_class == Lutaml::Model::TomlAdapter::TomlRbDocument
+    if adapter_class == Lutaml::Model::TomlAdapter::TomlRbAdapter
       TomlRB.dump(attributes)
-    else
+    elsif adapter_class == Lutaml::Model::TomlAdapter::TomlibAdapter
       Tomlib.dump(attributes)
     end
   end
@@ -23,22 +22,18 @@ RSpec.shared_examples "a TOML adapter" do |adapter_class|
   end
 
   it "deserializes from TOML" do
-    toml = if adapter_class == Lutaml::Model::TomlAdapter::TomlRbDocument
-             TomlRB.dump(attributes)
-           else
-             Tomlib.dump(attributes)
-           end
-    doc = adapter_class.parse(toml)
+    doc = adapter_class.parse(expected_toml)
     new_model = SampleModel.new(doc.to_h)
+
     expect(new_model.name).to eq("John Doe")
     expect(new_model.age).to eq(30)
   end
 end
 
-RSpec.describe Lutaml::Model::TomlAdapter::TomlRbDocument do
+RSpec.describe Lutaml::Model::TomlAdapter::TomlRbAdapter do
   it_behaves_like "a TOML adapter", described_class
 end
 
-RSpec.describe Lutaml::Model::TomlAdapter::TomlibDocument do
+RSpec.describe Lutaml::Model::TomlAdapter::TomlibAdapter do
   it_behaves_like "a TOML adapter", described_class
 end
