+++ spec/lutaml/model/xml_adapter/xml_namespace_spec.rb
@@ -1,4 +1,3 @@
-# spec/lutaml/model/xml_adapter_spec.rb
 require "spec_helper"
 require "lutaml/model/xml_adapter/nokogiri_adapter"
 require "lutaml/model/xml_adapter/ox_adapter"
@@ -25,7 +24,7 @@ class SampleNamespacedModel < Lutaml::Model::Serializable
   end
 end
 
-class NamespaceNill < Lutaml::Model::Serializable
+class NamespaceNil < Lutaml::Model::Serializable
   attribute :namespace_model, SampleNamespacedModel
 
   xml do
@@ -71,10 +70,10 @@ RSpec.shared_examples "an XML namespace parser" do |adapter_class|
 
     doc = adapter_class.parse(xml)
     new_model = SampleNamespacedModel.new(doc.root.children.to_h do |child|
-                                            [
-                                              child.unprefixed_name.downcase.to_sym, child.text
-                                            ]
-                                          end)
+      [
+        child.unprefixed_name.downcase.to_sym, child.text,
+      ]
+    end)
     expect(new_model.name).to eq("John Doe")
     expect(new_model.age).to eq(30)
   end
@@ -95,29 +94,29 @@ RSpec.shared_examples "an XML namespace parser" do |adapter_class|
 
     it "round-trips if namespace is set to nil in parent" do
       xml = <<~XML
-        <NamespaceNill xmlns:foo="http://example.com/foo" xmlns:bar="http://example.com/bar" xmlns:baz="http://example.com/baz">
+        <NamespaceNil xmlns:foo="http://example.com/foo" xmlns:bar="http://example.com/bar" xmlns:baz="http://example.com/baz">
           <SampleNamespacedModel xml:lang="en">
             <bar:Name>John Doe</bar:Name>
             <baz:Age>30</baz:Age>
           </SampleNamespacedModel>
-        </NamespaceNill>
+        </NamespaceNil>
       XML
 
-      doc = NamespaceNill.from_xml(xml)
+      doc = NamespaceNil.from_xml(xml)
       generated_xml = doc.to_xml
       expect(generated_xml).to be_equivalent_to(xml)
     end
   end
 end
 
-RSpec.describe Lutaml::Model::XmlAdapter::NokogiriDocument do
+RSpec.describe Lutaml::Model::XmlAdapter::NokogiriAdapter do
   it_behaves_like "an XML namespace parser", described_class
 end
 
-RSpec.describe Lutaml::Model::XmlAdapter::OxDocument do
+RSpec.describe Lutaml::Model::XmlAdapter::OxAdapter do
   it_behaves_like "an XML namespace parser", described_class
 end
 
-RSpec.xdescribe Lutaml::Model::XmlAdapter::OgaDocument do
+RSpec.xdescribe Lutaml::Model::XmlAdapter::OgaAdapter do
   it_behaves_like "an XML namespace parser", described_class
 end
