+++ spec/spec_helper.rb
@@ -20,15 +20,15 @@ require_relative "../lib/lutaml/model"
 require_relative "../lib/lutaml/model/xml_adapter/nokogiri_adapter"
 require_relative "../lib/lutaml/model/xml_adapter/ox_adapter"
 require_relative "../lib/lutaml/model/xml_adapter/oga_adapter"
-require_relative "../lib/lutaml/model/json_adapter/standard"
-require_relative "../lib/lutaml/model/json_adapter/multi_json"
-require_relative "../lib/lutaml/model/yaml_adapter"
+require_relative "../lib/lutaml/model/json_adapter/standard_json_adapter"
+require_relative "../lib/lutaml/model/json_adapter/multi_json_adapter"
+require_relative "../lib/lutaml/model/yaml_adapter/standard_yaml_adapter"
 require_relative "../lib/lutaml/model/toml_adapter/toml_rb_adapter"
 require_relative "../lib/lutaml/model/toml_adapter/tomlib_adapter"
 
 Lutaml::Model::Config.configure do |config|
-  config.xml_adapter = Lutaml::Model::XmlAdapter::NokogiriDocument
-  config.json_adapter = Lutaml::Model::JsonAdapter::StandardDocument
-  config.yaml_adapter = Lutaml::Model::YamlAdapter::Standard
-  config.toml_adapter = Lutaml::Model::TomlAdapter::TomlRbDocument
+  config.xml_adapter = Lutaml::Model::XmlAdapter::NokogiriAdapter
+  config.json_adapter = Lutaml::Model::JsonAdapter::StandardJsonAdapter
+  config.yaml_adapter = Lutaml::Model::YamlAdapter::StandardYamlAdapter
+  config.toml_adapter = Lutaml::Model::TomlAdapter::TomlRbAdapter
 end
